Fix prototypes for sudo_pread() and sudo_pwrite().

+++ include/sudo_compat.h
@@ -586,12 +586,12 @@ sudo_dso_public int sudo_pipe2(int fildes[2], int flags);
 # define pipe2(_a, _b) sudo_pipe2((_a), (_b))
 #endif /* HAVE_PIPE2 */
 #ifndef HAVE_PREAD
-sudo_dso_public int sudo_pread(int fd, void *buf, size_t nbytes, off_t offset);
+sudo_dso_public ssize_t sudo_pread(int fd, void *buf, size_t nbytes, off_t offset);
 # undef pread
 # define pread(_a, _b, _c, _d) sudo_pread((_a), (_b), (_c), (_d))
 #endif /* HAVE_PREAD */
 #ifndef HAVE_PWRITE
-sudo_dso_public int sudo_pwrite(int fd, const void *buf, size_t nbytes, off_t offset);
+sudo_dso_public ssize_t sudo_pwrite(int fd, const void *buf, size_t nbytes, off_t offset);
 # undef pwrite
 # define pwrite(_a, _b, _c, _d) sudo_pwrite((_a), (_b), (_c), (_d))
 #endif /* HAVE_PWRITE */
